# src/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Gestion UH API service for centralized user administration",
        "title": "Gestion UH API",
        "contact": {
            "email": "nodo@uh.cu"
        },
        "version": "1.0"
    },
    "host": "identity.sic.uh.cu",
    "paths": {
        "/admin/reset/workers": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Reset worker password",
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/admin/roles": {
            "get": {
                "description": "Get available roles.",
                "tags": [
                    "Admin"
                ],
                "summary": "Get available roles.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.RoleModel"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/admin/user": {
            "get": {
                "description": "GetUserInfo from user front the CI",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get User Info by CI.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CI of user to get info",
                        "name": "ci",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "UID of user to get info",
                        "name": "uid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.UserDetailsModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
...
        "/admin/user/changeEmail": {
            "post": {
                "description": "Change user email, only if the new email are available",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Change user email",
                "parameters": [
                    {
                        "description": "current email and the new email",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ChangeEmailData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/admin/user/changepassword": {
            "post": {
                "description": "Change user password stored in LDAP server.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Change user password by admin",
                "parameters": [
                    {
                        "description": "Account and new Password",
                        "name": "newPassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ForceChangePasswordModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/admin/user/createVPN": {
            "post": {
                "description": "Create VPN of given user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Create VPN of given user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CI of user to get VPN",
                        "name": "ci",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.VPNProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/admin/user/deleteVPN": {
            "post": {
                "description": "Delete VPN of given user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Delete VPN of given user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CI of user to get VPN",
                        "name": "ci",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/admin/user/disableVPN": {
            "post": {
                "description": "Disable VPN of given user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Disable VPN of given user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CI of user to disable VPN",
                        "name": "ci",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/admin/user/editServices": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Edit user services status",
                "parameters": [
                    {
                        "description": "patch data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.EditServicesStatusData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/admin/user/enableVPN": {
            "post": {
                "description": "Enable VPN of given user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Enable VPN of given user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CI of user to enable VPN",
                        "name": "ci",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/admin/user/reset": {
            "post": {
                "description": "Reset user and clear password",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Reset user password.",
                "parameters": [
                    {
                        "description": "CI and Email of user to reset",
                        "name": "reset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Reset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/admin/user/role": {
            "post": {
                "description": "Set role to user with id user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Set role to user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email of the user",
                        "name": "user",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "roleId to assign",
                        "name": "roleId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/admin/user/statusVPN": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Check if the use has a VPN created",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CI of user to check VPN",
                        "name": "ci",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StatusVPNResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ChangeEmailData": {
            "type": "object",
            "properties": {
                "currentEmail": {
                    "type": "string"
                },
                "newEmail": {
                    "type": "string"
                }
            }
        },
...
        "api.EditServicesStatusData": {
            "type": "object",
            "properties": {
                "hasCloud": {
                    "type": "boolean"
                },
                "hasEmail": {
                    "type": "boolean"
                },
                "hasInternet": {
                    "type": "boolean"
                },
                "hasVideoConference": {
                    "type": "boolean"
                },
                "hasVpn": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Error": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.ForceChangePasswordModel": {
            "type": "object",
            "required": [
                "email",
                "newPassword"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "newPassword": {
                    "type": "string"
                }
            }
        },
...
        "api.Reset": {
            "type": "object",
            "properties": {
                "ci": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
...
        "api.RoleModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "api.UserDetailsModel": {
            "type": "object",
            "properties": {
                "accountsDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.AdminUserDataModel"
                    }
                },
                "inSystem": {
                    "type": "boolean"
                }
            }
        },
        "api.VPNProfile": {
            "type": "object",
            "properties": {
                "vpnProfile": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}